+++ frontend/src/components/TopActivities.jsx
@@ -3,14 +3,12 @@ import MoreLessButton from "./molecules/MoreLessButton";
 
 function TopActivities({ data }) {
   return (
-    <div>
+    <div className="topActivitiesList">
       <div className="row border">
         {data.map((elem, index) => {
           return (
-            <div key={index} className="col-md-4 col-xl-2 col-xs-12">
-      
+            <div key={index} className="col">
                 <div className="border">in the last {elem[0]} hours</div>
-                <br />
                 {elem[1].map((elem, index) => (
                   <div style={{minHeight:"50px"}} className="border" key={index}>
                     <MoreLessButton
@@ -22,7 +20,7 @@ function TopActivities({ data }) {
                     />
                   </div>
                 ))}
-              </div>
+            </div>
            
           );
         })}