+++ frontend/src/components/atoms/LastUpdated.jsx
@@ -1,7 +1,6 @@
 import React from 'react';
-import Button from 'react-bootstrap/Button';
 
-function LastUpdated({time,setBgColor}) {
+function LastUpdated({time}) {
   const stickyStyle = {
     position: "sticky",
     top: 0,
@@ -10,11 +9,9 @@ function LastUpdated({time,setBgColor}) {
   }
   return (
     <div style={stickyStyle}>
-        <div style={{fontSize:"1vh"}}>
+        <div style={{fontSize:"2vh"}}>
             Last Updated at {time}
         </div>
-        <Button style={{fontSize:"1vh"}} size="sm" onClick={()=>setBgColor("white")} variant="light">Light</Button>
-        <Button style={{fontSize:"1vh"}} size="sm" onClick={()=>setBgColor("black")} variant="dark">dark</Button>
     </div>
   )
 }